# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      parameters:
        - name: scale
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/BusinessScale"
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /reports/{reportId}/data-points:
    get:
      operationId: listDataPoints
      tags: [dataPoints]
      summary: List data points for a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataPoint"
    post:
      operationId: createDataPoint
      tags: [dataPoints]
      summary: Add a data point to a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataPointInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataPoint"
  /data-points/{id}:
    patch:
      operationId: updateDataPoint
      tags: [dataPoints]
      summary: Update a data point
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataPointUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataPoint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteDataPoint
      tags: [dataPoints]
      summary: Delete a data point
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/{reportId}/constraints:
    get:
      operationId: listConstraints
      tags: [constraints]
      summary: List constraints for a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Constraints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Constraint"
    post:
      operationId: createConstraint
      tags: [constraints]
      summary: Add a constraint to a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstraintInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Constraint"
  /constraints/{id}:
    patch:
      operationId: updateConstraint
      tags: [constraints]
      summary: Update a constraint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstraintUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Constraint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteConstraint
      tags: [constraints]
      summary: Delete a constraint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/{reportId}/activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List audit trail entries for a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Add an audit trail entry
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BusinessScale:
      type: string
      enum: [mikro, kecil, menengah, besar]
      description: Investment scale determining reporting frequency
    OperatingMode:
      type: string
      enum: [pendamping, hibrida, penuh]
      description: Working mode — UMK companion, hybrid, or full human-in-loop
    PermitType:
      type: string
      enum: [nib, sertifikat_standar, izin]
      description: Jenis perizinan berusaha — NIB saja, NIB + Sertifikat Standar, atau NIB + Izin
    SsStatus:
      type: string
      enum: [tidak_ada, pernyataan_mandiri, perlu_verifikasi, terverifikasi]
      description: Status Sertifikat Standar (RBA) — tidak ada, pernyataan mandiri (MR), perlu verifikasi (MT), atau terverifikasi
...
    DataCategory:
      type: string
      enum: [investasi, tenaga_kerja, produksi, kewajiban]
      description: LKPM data component category
    DataStatus:
      type: string
      enum: [terverifikasi, perlu_verifikasi, estimasi]
      description: Verification status of a data point
...
    Company:
      type: object
      required:
        - id
        - name
        - nib
        - scale
        - operatingMode
        - permitType
        - ssStatus
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        nib:
          type: string
        scale:
          $ref: "#/components/schemas/BusinessScale"
        operatingMode:
          $ref: "#/components/schemas/OperatingMode"
        permitType:
          $ref: "#/components/schemas/PermitType"
        ssStatus:
          $ref: "#/components/schemas/SsStatus"
        kbli:
          type: array
          items:
            type: string
        capital:
          type: ["number", "null"]
          description: Modal usaha in IDR
        address:
          type: ["string", "null"]
        picName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CompanyInput:
      type: object
      required:
        - name
        - nib
        - scale
        - operatingMode
        - permitType
      properties:
        name:
          type: string
          minLength: 1
        nib:
          type: string
          minLength: 1
        scale:
          $ref: "#/components/schemas/BusinessScale"
        operatingMode:
          $ref: "#/components/schemas/OperatingMode"
        permitType:
          $ref: "#/components/schemas/PermitType"
        ssStatus:
          $ref: "#/components/schemas/SsStatus"
        kbli:
          type: array
          items:
            type: string
        capital:
          type: number
        address:
          type: string
        picName:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        nib:
          type: string
          minLength: 1
        scale:
          $ref: "#/components/schemas/BusinessScale"
        operatingMode:
          $ref: "#/components/schemas/OperatingMode"
        permitType:
          $ref: "#/components/schemas/PermitType"
        ssStatus:
          $ref: "#/components/schemas/SsStatus"
        kbli:
          type: array
          items:
            type: string
        capital:
          type: number
        address:
          type: string
        picName:
          type: string
...
    DataPoint:
      type: object
      required:
        - id
        - reportId
        - category
        - label
        - status
        - confidence
        - createdAt
      properties:
        id:
          type: integer
        reportId:
          type: integer
        category:
          $ref: "#/components/schemas/DataCategory"
        label:
          type: string
          description: e.g. "Realisasi modal tetap"
        fieldKey:
          type: ["string", "null"]
          description: >-
            Stable identifier binding this data point to an OSS catalog field
            (e.g. "inv_mesin"). When present, it is authoritative for
            catalog matching so label edits do not break the binding.
        value:
          type: ["number", "null"]
          description: Numeric realization value
        unit:
          type: ["string", "null"]
          description: e.g. "IDR", "orang"
        source:
          type: ["string", "null"]
          description: Provenance citation, e.g. "GL-2026-Q2, baris 88"
        status:
          $ref: "#/components/schemas/DataStatus"
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        attribution:
          type: ["string", "null"]
          description: Who/which agent produced this value
        createdAt:
          type: string
          format: date-time
    DataPointInput:
      type: object
      required:
        - category
        - label
        - status
        - confidence
      properties:
        category:
          $ref: "#/components/schemas/DataCategory"
        label:
          type: string
          minLength: 1
        fieldKey:
          type: string
          description: Stable OSS catalog field identifier (e.g. "inv_mesin").
        value:
          type: number
        unit:
          type: string
        source:
          type: string
        status:
          $ref: "#/components/schemas/DataStatus"
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        attribution:
          type: string
    DataPointUpdate:
      type: object
      properties:
        category:
          $ref: "#/components/schemas/DataCategory"
        label:
          type: string
          minLength: 1
        fieldKey:
          type: ["string", "null"]
          description: Stable OSS catalog field identifier (e.g. "inv_mesin").
        value:
          type: number
        unit:
          type: string
        source:
          type: string
        status:
          $ref: "#/components/schemas/DataStatus"
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        attribution:
          type: string
    Constraint:
      type: object
      required:
        - id
        - reportId
        - issue
        - createdAt
      properties:
        id:
          type: integer
        reportId:
          type: integer
        issue:
          type: string
        followUp:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ConstraintInput:
      type: object
      required:
        - issue
      properties:
        issue:
          type: string
          minLength: 1
        followUp:
          type: string
    ConstraintUpdate:
      type: object
      properties:
        issue:
          type: string
          minLength: 1
        followUp:
          type: string
    Activity:
      type: object
      required:
        - id
        - reportId
        - action
        - actor
        - createdAt
      properties:
        id:
          type: integer
        reportId:
          type: integer
        action:
          type: string
          description: e.g. "Gate G3 lolos", "Data divalidasi"
        actor:
          type: string
          description: Person or agent, e.g. "S3 · Validasi" or "Budi (Checker)"
        detail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ActivityInput:
      type: object
      required:
        - action
        - actor
      properties:
        action:
          type: string
          minLength: 1
        actor:
          type: string
          minLength: 1
        detail:
          type: string
...